pyproject.toml
[project]
name = "hackathon-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=22.0.0",
    "certifi>=2025.7.9",
    "easyocr>=1.7.2",
    "fastapi>=0.104.0",
    "flake8>=4.0.0",
    "huggingface-hub>=0.23.0",
    "jinja2>=3.1.0",
    "joblib>=1.1.0",
    "lap>=0.5.12",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "opencv-python>=4.8.0",
    "pandas>=1.3.0",
    "plotly>=5.0.0",
    "psutil>=5.8.0",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pyyaml>=6.0",
    "safetensors>=0.4.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "supervision>=0.22.0",
    "tensorboard>=2.8.0",
    "timm>=1.0.0",
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "tqdm>=4.64.0",
    "ultralytics>=8.3.165",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
]
